Tools: Migrated ScrollbarAxis out of Scrollbar.
